change uint64_t to uint32_t for #5590

+++ cores/esp8266/Esp.h
@@ -144,7 +144,7 @@ class EspClass {
 
         bool rtcUserMemoryRead(uint32_t offset, uint32_t *data, size_t size);
         bool rtcUserMemoryWrite(uint32_t offset, uint32_t *data, size_t size);
-	uint64_t getRtcTime(); // as microseconds
+	    uint32_t getRtcTime(); // as microseconds
 
         void reset();
         void restart();
